Rename infrastructure package name to infra
