+++ db/postgres.go
@@ -2,7 +2,7 @@ package db
 
 import (
 	"github.com/gghcode/apas-todo-apiserver/config"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/model"
+	"github.com/gghcode/apas-todo-apiserver/infra/model"
 	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
