+++ infra/file/afero_reader.go
