+++ infra/repository/todo_repository_test.go
@@ -6,8 +6,8 @@ import (
 	"github.com/gghcode/apas-todo-apiserver/config"
 	"github.com/gghcode/apas-todo-apiserver/db"
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/todo"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/model"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/repository"
+	"github.com/gghcode/apas-todo-apiserver/infra/model"
+	"github.com/gghcode/apas-todo-apiserver/infra/repository"
 	"github.com/gghcode/apas-todo-apiserver/internal/testutil"
 	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/suite"
