+++ infra/repository/token_repository_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gghcode/apas-todo-apiserver/config"
 	"github.com/gghcode/apas-todo-apiserver/db"
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/auth"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/repository"
+	"github.com/gghcode/apas-todo-apiserver/infra/repository"
 	"github.com/stretchr/testify/suite"
 )
