+++ wire.go
@@ -11,10 +11,10 @@ import (
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/auth"
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/todo"
 	"github.com/gghcode/apas-todo-apiserver/domain/usecase/user"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/file"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/jwt"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/repository"
-	"github.com/gghcode/apas-todo-apiserver/infrastructure/security"
+	"github.com/gghcode/apas-todo-apiserver/infra/file"
+	"github.com/gghcode/apas-todo-apiserver/infra/jwt"
+	"github.com/gghcode/apas-todo-apiserver/infra/repository"
+	"github.com/gghcode/apas-todo-apiserver/infra/security"
 	"github.com/gghcode/apas-todo-apiserver/web"
 	"github.com/gghcode/apas-todo-apiserver/web/api"
 	webApp "github.com/gghcode/apas-todo-apiserver/web/api/app"
